refactor: react-query -> fetcher 변경
