+++ src/fetcher/index.ts
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 import { backendFetch } from '@/fetcher/instance.ts';
-import { DefaultRes, WordDetail } from './types.ts';
+import type { DefaultRes, WordDetail, SearchWord } from './types.ts';
 
 export const getWordDetail = async (wordId: string) => {
   try {
@@ -8,6 +8,18 @@ export const getWordDetail = async (wordId: string) => {
   } catch (e) {
     // NOTE: 상황에 맞는 페이지 보여줘야 함.
     console.log('error');
+    notFound();
+  }
+};
+
+export const getWordSearch = async (wordName: string) => {
+  try {
+    return await backendFetch<DefaultRes<SearchWord>>(
+      `/word/search?keyword=${wordName}&page=1&limit=50`,
+    );
+  } catch (e) {
+    // NOTE: 상황에 맞는 페이지 보여줘야 함.
+    console.log('error', e);
     notFound();
   }
 };
